Signed current beta, generated signature for stable but did not publish it for routers
There is a user having strange issues in the IRC using version 0.6.4+exp20150224. Because I don't know what the difference between 0.6.4 and 0.6.4+exp20150224 is and because I do not understand the problem I do not want to push the updates myself. But this commit gives @dracoTrier everything he needs to push the update.