stable_canidate: remove tackin signature and set DATE to now+1 week
If we'd push out the firmware with the old value it would force all routers to update within an hour. That's not what we want. We want a slow rollout. I set the date 7 days into the future to give the other people 7 days to sign before release.

Removing tackins signature was nececarry because changing the date invalidated it and I can not recreate it myself.